게시글 수정시 내용 변경이 없다면 ai 답변 X

+++ src/main/java/com/credential/cubrism/server/posts/service/PostService.java
@@ -80,7 +80,7 @@ public class PostService {
 
         postRepository.save(post);
 
-        aiResponseService.updatePostWithAiResponse(post);
+        aiResponseService.postWithAiResponse(post);
 
         System.out.println("게시글 작성 완료");
         return ResponseEntity.status(HttpStatus.CREATED).body(new MessageDto("게시글을 작성했습니다."));
@@ -125,6 +125,8 @@ public class PostService {
             throw new CustomException(ErrorCode.UPDATE_DENIED);
         }
 
+        String oldContent = post.getContent();
+
         // 삭제할 이미지가 존재하면 S3 및 DB에서 이미지 삭제
         if (!dto.getRemovedImages().isEmpty()) {
             dto.getRemovedImages().forEach(imageUrl -> {
@@ -154,14 +156,18 @@ public class PostService {
         post.setTitle(dto.getTitle());
         post.setContent(dto.getContent());
         post.setQualificationList(qualificationList);
-        if (post.getPostAiComments() != null) {
-            postAiCommentsRepository.delete(post.getPostAiComments());
-            post.setPostAiComments(null);
+
+        if(!(oldContent.equals(post.getContent()))){
+            System.out.println("내용이 변경되었습니다");
+            if (post.getPostAiComments() != null) {
+                postAiCommentsRepository.delete(post.getPostAiComments());
+                post.setPostAiComments(null);
+            }
+            aiResponseService.postWithAiResponse(post);
+        } else{
+            System.out.println("내용이 변경되지 않았습니다");
         }
-
         postRepository.save(post);
-
-        aiResponseService.updatePostWithAiResponse(post);
 
         System.out.println("게시글 수정 완료");
         return ResponseEntity.status(HttpStatus.CREATED).body(new MessageDto("게시글을 수정했습니다."));
